# docs/product-kit/04_data_api/openapi.yaml
openapi: 3.1.0
info:
  title: Work Calendar API
  version: 1.0.0
  description: 工作日历小程序、APP 与管理后台统一接口草案。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /schedules:
    get:
      tags: [Schedules]
      summary: 查询排班实例
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - name: ownerUserId
          in: query
          schema: {type: string, format: uuid}
        - name: from
          in: query
          required: true
          schema: {type: string, format: date}
        - name: to
          in: query
          required: true
          schema: {type: string, format: date}
      responses:
        '200':
          description: 排班列表
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ScheduleInstance'}
    post:
      tags: [Schedules]
      summary: 创建单日排班
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - {$ref: '#/components/parameters/IdempotencyHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ScheduleCreateInput'}
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScheduleInstance'}
        '409': {$ref: '#/components/responses/Conflict'}
  /schedules/{id}:
    get:
      tags: [Schedules]
      summary: 获取排班详情
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - {$ref: '#/components/parameters/ResourceId'}
      responses:
        '200':
          description: 排班详情
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScheduleDetail'}
    patch:
      tags: [Schedules]
      summary: 修改排班实例
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - {$ref: '#/components/parameters/ResourceId'}
        - {$ref: '#/components/parameters/IdempotencyHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version, changeScope]
              properties:
                version: {type: integer, minimum: 1}
                changeScope: {type: string, enum: [only_this_day, future_instances, whole_rule]}
                shiftTemplateId: {type: string, format: uuid}
                customShift: {$ref: '#/components/schemas/CustomShift'}
                locationId: {type: [string, 'null'], format: uuid}
                note: {type: [string, 'null'], maxLength: 1000}
                reason: {type: string, maxLength: 500}
      responses:
        '200':
          description: 已修改
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ScheduleInstance'}
        '409': {$ref: '#/components/responses/Conflict'}
  /schedule-rules:
    post:
      tags: [Schedules]
      summary: 创建循环排班规则
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - {$ref: '#/components/parameters/IdempotencyHeader'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ScheduleRuleInput'}
      responses:
        '201':
          description: 已创建并生成排班
          content:
            application/json:
              schema:
                type: object
                properties:
                  rule: {$ref: '#/components/schemas/ScheduleRule'}
                  generatedCount: {type: integer}
                  conflicts:
                    type: array
                    items: {$ref: '#/components/schemas/ScheduleConflict'}
    get:
      tags: [Schedules]
      summary: 列出当前用户的循环排班表
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
      responses:
        '200':
          description: 排班表列表
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: {type: string, format: uuid}
                    name: {type: string}
                    startDate: {type: string, format: date}
                    endDate: {type: [string, 'null'], format: date}
                    timezone: {type: string}
                    isCurrent: {type: boolean}
  /schedule-rules/{id}/activate:
    post:
      tags: [Schedules]
      summary: 切换当前排班表
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - {$ref: '#/components/parameters/ResourceId'}
      responses:
        '200':
          description: 已切换
          content:
            application/json:
              schema:
                type: object
                properties:
                  ruleId: {type: string, format: uuid}
  /schedule-rules/{id}:
    patch:
      tags: [Schedules]
      summary: 编辑排班表（名称/起始/结束/序列）
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - {$ref: '#/components/parameters/ResourceId'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: {type: integer}
                name: {type: [string, 'null']}
                startDate: {type: string, format: date}
                endDate: {type: [string, 'null'], format: date}
                sequence:
                  type: array
                  items:
                    type: object
                    properties:
                      shiftTemplateId: {type: string, format: uuid}
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: {type: string, format: uuid}
                  name: {type: string}
                  startDate: {type: string, format: date}
                  endDate: {type: [string, 'null'], format: date}
                  isCurrent: {type: boolean}
    delete:
      tags: [Schedules]
      summary: 删除排班表及其生成的排班
      parameters:
        - {$ref: '#/components/parameters/WorkspaceHeader'}
        - {$ref: '#/components/parameters/ResourceId'}
      responses:
        '200':
          description: 已删除
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed: {type: string, format: uuid}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceHeader:
      name: X-Workspace-Id
      in: header
      required: true
      schema: {type: string, format: uuid}
    IdempotencyHeader:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 8, maxLength: 128}
    ResourceId:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
  responses:
    Conflict:
      description: 版本或排班冲突
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
  schemas:
...
    CustomShift:
      type: object
      required: [name, kind]
      properties:
        name: {type: string}
        kind: {type: string, enum: [work, rest, leave, training, travel, custom]}
        startTime: {type: [string, 'null']}
        endTime: {type: [string, 'null']}
        endsNextDay: {type: boolean}
        color: {type: string}
    ScheduleCreateInput:
      type: object
      required: [ownerUserId, businessDate]
      properties:
        ownerUserId: {type: string, format: uuid}
        businessDate: {type: string, format: date}
        shiftTemplateId: {type: string, format: uuid}
        customShift: {$ref: '#/components/schemas/CustomShift'}
        locationId: {type: [string, 'null'], format: uuid}
        note: {type: [string, 'null'], maxLength: 1000}
    ScheduleInstance:
      type: object
      required: [id, businessDate, kind, status, version, shiftSnapshot]
      properties:
        id: {type: string, format: uuid}
        ownerUserId: {type: string, format: uuid}
        businessDate: {type: string, format: date}
        timezone: {type: string}
        startsAt: {type: [string, 'null'], format: date-time}
        endsAt: {type: [string, 'null'], format: date-time}
        kind: {type: string, enum: [work, rest, leave, training, travel, custom]}
        status: {type: string, enum: [scheduled, pending_approval, cancelled, completed]}
        source: {type: string}
        shiftSnapshot: {type: object, additionalProperties: true}
        locationSnapshot: {type: [object, 'null'], additionalProperties: true}
        note: {type: [string, 'null']}
        version: {type: integer}
    ScheduleDetail:
      allOf:
        - {$ref: '#/components/schemas/ScheduleInstance'}
        - type: object
          properties:
            weather: {$ref: '#/components/schemas/WeatherForecast'}
            pendingChange: {$ref: '#/components/schemas/ChangeRequest'}
            history:
              type: array
              items: {type: object, additionalProperties: true}
    ScheduleRuleInput:
      type: object
      required: [ownerUserId, startDate, sequence, timezone]
      properties:
        ownerUserId: {type: string, format: uuid}
        name: {type: [string, 'null']}
        startDate: {type: string, format: date}
        endDate: {type: [string, 'null'], format: date}
        timezone: {type: string}
        sequence:
          type: array
          minItems: 1
          items:
            type: object
            required: [shiftTemplateId]
            properties:
              shiftTemplateId: {type: string, format: uuid}
        generationHorizonDays: {type: integer, minimum: 7, maximum: 366, default: 90}
    ScheduleRule:
      allOf:
        - {$ref: '#/components/schemas/ScheduleRuleInput'}
        - type: object
          properties:
            id: {type: string, format: uuid}
            version: {type: integer}
            isActive: {type: boolean}
    ScheduleConflict:
      type: object
      required: [type, severity, message]
      properties:
        type: {type: string, enum: [overlap, insufficient_rest, coverage_gap, policy]}
        severity: {type: string, enum: [warning, error]}
        message: {type: string}
        existingScheduleId: {type: [string, 'null'], format: uuid}
...
    ChangeRequest:
      type: object
      required: [id, status, originalSnapshot, requestedSnapshot, createdAt]
      properties:
        id: {type: string, format: uuid}
        scheduleInstanceId: {type: string, format: uuid}
        status: {type: string, enum: [pending, approved, rejected, withdrawn, expired]}
        originalSnapshot: {type: object, additionalProperties: true}
        requestedSnapshot: {type: object, additionalProperties: true}
        reason: {type: [string, 'null']}
        approvalNote: {type: [string, 'null']}
        createdAt: {type: string, format: date-time}
        decidedAt: {type: [string, 'null'], format: date-time}
    WeatherForecast:
      type: object
      properties:
        date: {type: string, format: date}
        conditionCode: {type: string}
        conditionText: {type: string}
        temperatureMin: {type: number}
        temperatureMax: {type: number}
        humidityPercent: {type: integer}
        precipitationProbability: {type: integer}
        windDirection: {type: string}
        windLevel: {type: string}
        airQuality: {type: string}
        warningCodes:
          type: array
          items: {type: string}
        updatedAt: {type: string, format: date-time}
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: {type: string}
            message: {type: string}
            requestId: {type: string}
            details: {type: [object, 'null'], additionalProperties: true}